Force $SHELL to /bin/bash during install
When I did the install, the .perlbrew/init file used `setenv` commands to setup the environment variables.  That command is apparently a tsch equivilent command to export and results in:

-bash: setenv: command not found

Reading through https://github.com/gugod/App-perlbrew/issues/221 , it appears that the installer checks the value of $SHELL to determine which syntax to use.  Add the environment variable to the exec.  This seems to fix it.